Use std::ssize instead of casting, fix a lot of warnings and simplify code

+++ src/plugins/score-plugin-scenario/Scenario/Inspector/Expression/SimpleExpressionEditorWidget.cpp
@@ -73,7 +73,7 @@ public:
 
 SimpleExpressionEditorWidget::SimpleExpressionEditorWidget(
     const score::DocumentContext& doc,
-    int index,
+    int64_t index,
     QWidget* parent,
     QMenu* menu)
     : QWidget(parent)
